# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Image Storer",
        "contact": {}
    },
    "paths": {
        "/image": {
            "post": {
                "summary": "uploads new image to server",
                "operationId": "UploadImage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username of the user that wants to upload image",
                        "name": "username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "the file that the user wants to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "the filename of the corresponding file that the user wants to upload",
                        "name": "filename",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/image/:id": {
            "delete": {
                "summary": "deletes existing image of a user",
                "operationId": "DeleteImage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username of the user that wants to delete his/her image",
                        "name": "username",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "summary": "renames existing image of a user",
                "operationId": "RenameImage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username of the user that wants to rename his/her image",
                        "name": "username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "new file name",
                        "name": "RenameFile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RenameFile"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/images": {
            "get": {
                "summary": "lists images that belong to the user",
                "operationId": "ListImages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username of the user that wants to list his/her image(s)",
                        "name": "username",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/login": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "summary": "logs user in",
                "operationId": "Login",
                "parameters": [
                    {
                        "description": "the user to be logged in",
                        "name": "User",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.User"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/signup": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "summary": "registers new user",
                "operationId": "SignUp",
                "parameters": [
                    {
                        "description": "new user",
                        "name": "User",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.User"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "dto.RenameFile": {
            "type": "object",
            "properties": {
                "new": {
                    "type": "string"
                }
            }
        },
        "dto.User": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}